pyserver/requirements.txt
Django>=5.2,<6.0
# API and auth
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
# CORS
django-cors-headers>=4.4,<5.0
# Static files in prod
whitenoise>=6.8,<7.0
# Optional Postgres URL parsing
psycopg[binary,pool]>=3.2,<4.0
# Optional: parse DATABASE_URL
dj-database-url>=2.2,<3.0
# Optional Redis cache
django-redis>=5.4,<6.0
gunicorn>=22.0,<23.0

# OpenAI SDK for server-side AI proxy
openai>=1.40,<2.0
